fix like unlike button in view.tpl

+++ likedeg.php
@@ -3,12 +3,13 @@
 include("include/config.php");
 include("include/functions/import.php");
 
-$pid = intval(cleanit($_REQUEST['pid']));
-$art = cleanit($_REQUEST['art']);
+$love = cleanit($_REQUEST['l']);
+$unlove = cleanit($_REQUEST['u']);
+$pid = cleanit($_REQUEST['pid']);
 $SID = intval($_SESSION['USERID']);
 if(($SID > 0) && ($pid > 0))
 {
-	if($art == "1")
+	if($love == "1")
 	{
 		$query="INSERT INTO posts_favorited SET PID='".mysql_real_escape_string($pid)."', USERID='".mysql_real_escape_string($SID)."'";
 		$result=$conn->execute($query);
@@ -21,40 +22,46 @@ if(($SID > 0) && ($pid > 0))
 			$expire=time()+60*60*24;
 			setcookie($cname, "yes", $expire);
 		}
+		$query="DELETE FROM posts_unfavorited WHERE PID='".mysql_real_escape_string($pid)."' AND USERID='".mysql_real_escape_string($SID)."'";
+		$result=$conn->execute($query);
 	}
-	elseif($art == "-1")
+	elseif($love == "-1")
 	{
 		$query="DELETE FROM posts_favorited WHERE PID='".mysql_real_escape_string($pid)."' AND USERID='".mysql_real_escape_string($SID)."'";
 		$result=$conn->execute($query);
 	}
-
-	$query="SELECT phase,favclicks,unfavclicks FROM posts WHERE PID='".mysql_real_escape_string($pid)."'";
-	$executequery=$conn->execute($query);
-	$phase = $executequery->fields['phase'];
-	$favclicks = $executequery->fields['favclicks'];
-	$unfavclicks = $executequery->fields['unfavclicks'];
-	if($phase == "0")
+	if($unlove == "1")
 	{
-		$myes = $config['myes'];
-		$mno = $config['mno'];
-		if($favclicks >= $myes)
+		$query="INSERT INTO posts_unfavorited SET PID='".mysql_real_escape_string($pid)."', USERID='".mysql_real_escape_string($SID)."'";
+		$result=$conn->execute($query);
+		$cname = "D".$pid;
+		if($pid > 0 && !isset($_COOKIE[$cname]))
 		{
-			$query="UPDATE posts SET phase='1', ttime='".time()."' WHERE PID='".mysql_real_escape_string($pid)."' AND phase='0'";
+			$query="UPDATE posts SET unfavclicks=unfavclicks+1 WHERE PID='".mysql_real_escape_string($pid)."'";
 			$result=$conn->execute($query);
+			$expire=time()+60*60*24;
+			setcookie($cname, "yes", $expire);
 		}
+		$query="DELETE FROM posts_favorited WHERE PID='".mysql_real_escape_string($pid)."' AND USERID='".mysql_real_escape_string($SID)."'";
+		$result=$conn->execute($query);
 	}
-	elseif($phase == "1")
+	elseif($unlove == "-1")
 	{
-		$mtrend = $config['mtrend'];
-		if($favclicks >= $mtrend)
-		{
-			$query="UPDATE posts SET phase='2', htime='".time()."' WHERE PID='".mysql_real_escape_string($pid)."' AND phase='1'";
-			$result=$conn->execute($query);
-			$query="DELETE FROM posts_unfavorited WHERE PID='".mysql_real_escape_string($pid)."'";
-			$result=$conn->execute($query);
-		}
+		$query="DELETE FROM posts_unfavorited WHERE PID='".mysql_real_escape_string($pid)."' AND USERID='".mysql_real_escape_string($SID)."'";
+		$result=$conn->execute($query);
 	}
-
+	$query="SELECT favclicks FROM posts WHERE PID='".mysql_real_escape_string($pid)."'";
+    $executequery=$conn->execute($query);
+    $favclicks = $executequery->fields['favclicks'];
+	$mtrend = $config['mtrend'];
+	if($favclicks >= $mtrend)
+	{
+		$query="UPDATE posts SET phase='2', htime='".time()."' WHERE PID='".mysql_real_escape_string($pid)."' AND phase='1'";
+		$result=$conn->execute($query);
+		$query="DELETE FROM posts_unfavorited WHERE PID='".mysql_real_escape_string($pid)."'";
+		$result=$conn->execute($query);
+	}
+		
 	$query="SELECT count(*) as total FROM posts_favorited WHERE PID='".mysql_real_escape_string($pid)."'";
 	$executequery=$conn->execute($query);
 	$total = $executequery->fields[total];
